Provide several modes for displaying list of added/mod/untrack files
file_list_mode=fileonly
Displays just the filename only regardless of which directory a changed
file is at.

file_list_mode=fullpath
Displays the path and filename of changed files

file_list_mode=topdir
Display just the top level directories if there are changes in that
directory or any sub-directory and filenames if they are in the base
directory of the repo.
This is the behavior that exists before this commit

file_list_mode=fulldir
Display the full directory name for directories where there are changes
and filenames if they are in the base directory of the repo.
IMHO, this is a more useful mode.

Signed-off-by: Jon Ringle <jringle@gridpoint.com>